Added Image and Video Navigation
